requirements.txt
Django==1.8.4
Pillow==2.9.0
Werkzeug==0.10.4
django-appconf==1.0.1
django-bootstrap3==6.2.2
django-extensions==1.5.6
django-imagekit==3.2.7
django-taggit==0.17.1
pilkit==1.1.13
six==1.9.0
